refactor: move mockData to separate file
